+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp.hs
@@ -61,9 +61,7 @@ import Test.Cardano.Ledger.Shelley.ImpTest (ImpInit)
 
 spec ::
   forall era.
-  ( Arbitrary (TxAuxData era)
-  , ConwayEraImp era
-  , EraSegWits era
+  ( ConwayEraImp era
   , Inject (BabbageContextError era) (ContextError era)
   , Inject (ConwayContextError era) (ContextError era)
   , InjectRuleFailure "LEDGER" ConwayGovPredFailure era
@@ -103,7 +101,6 @@ spec = do
 conwaySpec ::
   forall era.
   ( ConwayEraImp era
-  , EraSegWits era
   , Inject (BabbageContextError era) (ContextError era)
   , Inject (ConwayContextError era) (ContextError era)
   , InjectRuleFailure "LEDGER" ConwayGovPredFailure era